+++ frontend/packs/request_to_static.jsx
@@ -1,0 +1,4 @@
+import RequestToStatic from "components_react/request_to_static/request_to_static";
+import WebpackerReact from "webpacker-react";
+
+WebpackerReact.setup({RequestToStatic});