+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - navigation icons
 - support in telegram
+- request to static at static page
 
 ### Modified
 - delete icon